move resumption testing and gauntlet to a separate test file
